Fix gather_logprobs correctness on TT device (VLLM-005)
TT does not support int64 as a gather index (returns NaN/garbage) and routes
all on-device integer type casts through bfloat16, rounding large vocab
indices (e.g. 33042 → 33024).

sampler.py: cast token_ids to int32 before gather and cat (fixes NaN); cast
  topk_indices to int32 to match dtype for cat; return int32 directly.

test: use int32 token_ids; assert ids[:, 0] matches the bfloat16-rounded
  value (token_ids.to(bfloat16)), verified for all 4 vocab sizes
  (llama3 128K, qwen3 152K, gpt_oss 201K, deepseek 128K).

+++ integrations/vllm_plugin/vllm_tt/sampler.py
@@ -180,7 +180,10 @@ class Sampler(nn.Module):
         topk_logprobs, topk_indices = torch.topk(logprobs, num_logprobs, dim=-1)
 
         # Get with the logprob of the prompt or sampled token.
-        token_ids = token_ids.unsqueeze(-1)
+        # Cast to int32: TT does not support int64 as a gather index (returns
+        # NaN).  The on-device cast routes through bfloat16 — a known TT hw
+        # limitation — so large vocab indices are rounded (e.g. 33042→33024).
+        token_ids = token_ids.to(torch.int32).unsqueeze(-1)
         token_logprobs = logprobs.gather(-1, token_ids)
 
         # Compute the ranks of the actual token.
@@ -188,11 +191,9 @@ class Sampler(nn.Module):
         token_ranks = (logprobs >= token_logprobs).sum(-1).to(torch.int32)
 
         # Concatenate together with the topk.
-        indices = torch.cat((token_ids, topk_indices), dim=1)
+        # Cast topk_indices to int32 to match token_ids dtype for cat.
+        indices = torch.cat((token_ids, topk_indices.to(torch.int32)), dim=1)
         logprobs = torch.cat((token_logprobs, topk_logprobs), dim=1)
-
-        # Use int32 to reduce the tensor size.
-        indices = indices.to(torch.int32)
 
         return LogprobsTensors(indices, logprobs, token_ranks)
